requirements.txt
fastapi==0.115.7
requests==2.32.3
urllib3==2.3.0
pillow==11.1.0
fpdf==1.7.2
Jinja2==3.1.2